Replace HighlightDriver with syntect::io::HighlightedWriter
HighlightedWriter is an implementation of std::io::Write that line-buffers
incoming bytes, validates UTF-8 at line boundaries, and forwards complete
lines through the existing parsing + ScopeRenderer pipeline. The previous
HighlightDriver type is gone; src/easy.rs is restored to its master state
modulo #[deprecated(since = "6.0.0")] annotations on HighlightLines and
HighlightFile.

The ScopeRenderer trait, render_line, and ThemedANSIScopeRenderer now live
in the new syntect::io module alongside HighlightedWriter. ClassedHTMLGenerator
keeps its line-oriented public surface and now wraps HighlightedWriter via
write_all. The HTML convenience helpers and the syncat / parsyncat / latex-demo
examples all switch to streaming via std::io::copy or write_all, dropping
their bespoke read_line loops.
