+++ examples/parsyncat.rs
@@ -2,12 +2,12 @@
 //! Prints the highlighted output to stdout.
 
 use rayon::prelude::*;
-use syntect::easy::HighlightDriver;
 use syntect::highlighting::ThemeSet;
+use syntect::io::HighlightedWriter;
 use syntect::parsing::SyntaxSet;
 
 use std::fs::File;
-use std::io::{self, BufRead, BufReader, Write};
+use std::io::{self, Write};
 
 fn main() {
     let files: Vec<String> = std::env::args().skip(1).collect();
@@ -20,40 +20,20 @@ fn main() {
     let syntax_set = SyntaxSet::load_defaults_newlines();
     let theme_set = ThemeSet::load_defaults();
 
-    // We first collect the contents of the files...
-    let contents: Vec<Vec<String>> = files
-        .par_iter()
-        .map(|filename| {
-            let mut lines = Vec::new();
-            // We use `String::new()` and `read_line()` instead of `BufRead::lines()`
-            // in order to preserve the newlines and get better highlighting.
-            let mut line = String::new();
-            let mut reader = BufReader::new(File::open(filename).unwrap());
-            while reader.read_line(&mut line).unwrap() > 0 {
-                lines.push(line);
-                line = String::new();
-            }
-            lines
-        })
-        .collect();
-
-    // ...then highlight each file in parallel, collecting rendered output...
+    // Highlight each file in parallel, buffering each result into its own
+    // `Vec<u8>` so we can stream them out in order at the end.
     let outputs: Vec<Vec<u8>> = files
         .par_iter()
-        .zip(&contents)
-        .map(|(filename, contents)| {
+        .map(|filename| {
             let theme = &theme_set.themes["base16-ocean.dark"];
             let syntax = syntax_set
                 .find_syntax_for_file(filename)
                 .unwrap()
                 .unwrap_or_else(|| syntax_set.find_syntax_plain_text());
-            let mut h = HighlightDriver::new(syntax, &syntax_set, theme);
-
-            for line in contents {
-                h.highlight_line(line).unwrap();
-            }
-
-            h.finalize()
+            let mut f = File::open(filename).unwrap();
+            let mut w = HighlightedWriter::new(syntax, &syntax_set, theme);
+            io::copy(&mut f, &mut w).unwrap();
+            w.finalize().unwrap()
         })
         .collect();
